STL-4796 Change dlid size in struct ib_ah_attr from 16 to 32 bits
OPA devices can support larger number of nodes than 16 bit lids
can accommodate. This change increases the dlid size from 16 to
32 bits in the kernel so kernel clients such as ib_sa can set up
address handles with 32 with dlid.
UAPI is unchanged and is still only 16 bits.

+++ hfi2/verbs/ruc.c
@@ -319,7 +319,7 @@ void hfi2_make_ruc_header(struct hfi2_qp *qp, struct ib_l4_headers *ohdr,
 	if (qp->s_mig_state == IB_MIG_MIGRATED)
 		bth0 |= IB_BTH_MIG_REQ;
 	qp->s_hdr->ph.ibh.lrh[0] = cpu_to_be16(lrh0);
-	qp->s_hdr->ph.ibh.lrh[1] = cpu_to_be16(qp->remote_ah_attr.dlid);
+	qp->s_hdr->ph.ibh.lrh[1] = cpu_to_be16((u16)qp->remote_ah_attr.dlid);
 	qp->s_hdr->ph.ibh.lrh[2] =
 		cpu_to_be16(qp->s_hdrwords + nwords);
 	qp->s_hdr->ph.ibh.lrh[3] = cpu_to_be16(ppd->lid |
